Determine whether or not an SDL3 monitor ID is valid using SDL_GetDisplayProperties()
On the SDL3 backend, when a laptop is connected to an external monitor and then the laptop lid is closed, the original SDL3 monitor ID for the laptop's monitor is invalidated and replaced with a monitor ID that exceeds the total amount of monitors upon reopening the laptop lid. Many of Raylib's functions in the file "rcore_desktop_sdl.c" do not properly handle such a situation, as they return an error message and do not actually run when the following condition is met:

#if defined(USING_VERSION_SDL3)
if((monitor > 0) && (monitor <= monitorCount))

I think the monitor number ("monitor") should be validated not against the monitor count, but rather the code should attempt to call the SDL3 function "SDL_GetDisplayProperties(monitor)", passing in the monitor id ("monitor"). This function returns a valid property ID on success or 0 on failure, so a returned value other than zero tells you that the monitor id is valid.

see this link for more details: https://wiki.libsdl.org/SDL3/SDL_GetDisplayProperties